+++ app/src/main/AndroidManifest.xml
@@ -11,17 +11,17 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity"
+        <activity android:name=".ui.MainActivity"
                   android:windowSoftInputMode="adjustNothing"/>
         <activity android:name=".recipe.RecipeActivity">
         </activity>
 
-        <activity android:name=".write.WriteActivity"/>
+        <activity android:name=".ui.write.WriteActivity"/>
 
-        <activity android:name=".search.SearchActivity"
+        <activity android:name=".ui.search.SearchActivity"
                   android:windowSoftInputMode="stateVisible|adjustNothing"/>
 
-        <activity android:name=".splash.SplashActivity"
+        <activity android:name=".ui.splash.SplashActivity"
                   android:theme="@style/Theme.AppCompat.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
@@ -29,6 +29,9 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+
+        <activity android:name=".ui.login.LoginActivity"
+                  android:theme="@style/Theme.AppCompat.NoActionBar"/>
     </application>
 
 </manifest>
