+++ app/src/main/java/cooking/depromeet/github/com/bangguseokcookking_android/ui/write/WriteRecipe.kt
@@ -1,0 +1,7 @@
+package cooking.depromeet.github.com.bangguseokcookking_android.ui.write
+
+data class WriteRecipe(
+    var image: String,
+    var content: String,
+    var number: Int
+)
