+++ app/src/main/java/cooking/depromeet/github/com/bangguseokcookking_android/ui/write/WriteRecipeEventListener.kt
@@ -1,0 +1,8 @@
+package cooking.depromeet.github.com.bangguseokcookking_android.ui.write
+
+interface WriteRecipeEventListener {
+
+    fun onImageSelected(position: Int)
+
+    fun onTextWatcher(text: String, position: Int)
+}
