--- app/src/main/java/cooking/depromeet/github/com/bangguseokcookking_android/write/WriteRecipe.kt
@@ -1,7 +1,0 @@
-package cooking.depromeet.github.com.bangguseokcookking_android.write
-
-data class WriteRecipe(
-    var image: String,
-    var content: String,
-    var number: Int
-)
