--- app/src/main/java/cooking/depromeet/github/com/bangguseokcookking_android/write/WriteRecipeEventListener.kt
@@ -1,8 +1,0 @@
-package cooking.depromeet.github.com.bangguseokcookking_android.write
-
-interface WriteRecipeEventListener {
-
-    fun onImageSelected(position: Int)
-
-    fun onTextWatcher(text: String, position: Int)
-}
