+++ app/src/main/res/drawable/login_et_focus.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_focused="true" android:color="#ffb022"/>
+
+    <item android:state_focused="false" android:color="#9b9b9b"/>
+
+</selector>
