+++ app/src/main/res/layout/activity_login.xml
@@ -1,0 +1,157 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:layout_width="match_parent"
+        android:background="@drawable/img_login"
+        android:layout_height="match_parent">
+
+    <FrameLayout
+            android:layout_width="0dp"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            android:background="#66000000"
+            android:layout_height="0dp"/>
+
+    <androidx.constraintlayout.widget.Guideline
+            android:layout_width="0dp"
+            android:orientation="vertical"
+            android:id="@+id/login_start_guide"
+            app:layout_constraintGuide_percent="0.056"
+            android:layout_height="0dp"/>
+
+    <androidx.constraintlayout.widget.Guideline
+            android:layout_width="0dp"
+            android:orientation="vertical"
+            android:id="@+id/login_end_guide"
+            app:layout_constraintGuide_percent="0.944"
+            android:layout_height="0dp"/>
+
+    <androidx.constraintlayout.widget.Guideline
+            android:layout_width="0dp"
+            android:orientation="horizontal"
+            android:id="@+id/login_top_guide"
+            app:layout_constraintGuide_percent="0.128"
+            android:layout_height="0dp"/>
+
+    <androidx.appcompat.widget.AppCompatTextView
+            app:layout_constraintTop_toBottomOf="@id/login_top_guide"
+            app:layout_constraintStart_toEndOf="@id/login_start_guide"
+            android:textSize="24sp"
+            android:textColor="#ffffff"
+            android:text="로그인"
+            android:id="@+id/login_tv_top"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"/>
+
+    <androidx.appcompat.widget.AppCompatTextView
+            android:id="@+id/login_tv_login"
+            android:layout_width="wrap_content"
+            android:textSize="18sp"
+            android:textColor="#ffffff"
+            android:text="아이디"
+            app:layout_constraintTop_toBottomOf="@id/login_tv_top"
+            app:layout_constraintStart_toEndOf="@id/login_start_guide"
+            android:layout_marginTop="50dp"
+            android:layout_height="wrap_content"/>
+
+    <androidx.appcompat.widget.AppCompatEditText
+            android:id="@+id/login_et_login"
+            android:layout_width="0dp"
+            app:layout_constraintTop_toBottomOf="@id/login_tv_login"
+            app:layout_constraintStart_toEndOf="@id/login_start_guide"
+            app:layout_constraintEnd_toStartOf="@id/login_end_guide"
+            android:backgroundTint="@drawable/login_et_focus"
+            android:layout_height="35dp"/>
+
+    <androidx.appcompat.widget.AppCompatTextView
+            android:id="@+id/login_tv_password"
+            android:textSize="18sp"
+            android:textColor="#ffffff"
+            android:text="비밀번호"
+            app:layout_constraintTop_toBottomOf="@id/login_et_login"
+            android:layout_marginTop="16dp"
+            app:layout_constraintStart_toEndOf="@id/login_start_guide"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"/>
+
+
+    <androidx.appcompat.widget.AppCompatEditText
+            android:id="@+id/login_et_password"
+            android:layout_width="0dp"
+            app:layout_constraintTop_toBottomOf="@id/login_tv_password"
+            app:layout_constraintStart_toEndOf="@id/login_start_guide"
+            app:layout_constraintEnd_toStartOf="@id/login_end_guide"
+            android:backgroundTint="@drawable/login_et_focus"
+            android:layout_height="35dp"/>
+
+    <androidx.constraintlayout.widget.Guideline
+            android:layout_width="0dp"
+            android:id="@+id/login_btn_top_guide"
+            android:orientation="horizontal"
+            app:layout_constraintGuide_percent="0.608"
+            android:layout_height="0dp"/>
+
+    <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/login_btn_kakao"
+            android:layout_width="0dp"
+            android:textSize="14sp"
+            android:textColor="#ffffff"
+            android:stateListAnimator="@null"
+            android:drawableStart="@drawable/ic_kakao"
+            android:paddingStart="27dp"
+            android:text="카카오톡으로 로그인하기"
+            android:paddingEnd="27dp"
+            android:gravity="center"
+            app:layout_constraintTop_toBottomOf="@id/login_btn_top_guide"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintWidth_default="percent"
+            android:background="@drawable/btn_login"
+            app:layout_constraintWidth_percent="0.631"
+            android:layout_height="48dp"/>
+
+    <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/login_btn_facebook"
+            android:layout_width="0dp"
+            android:textSize="14sp"
+            android:textColor="#ffffff"
+            android:stateListAnimator="@null"
+            android:drawableStart="@drawable/ic_facebook"
+            android:paddingStart="27dp"
+            android:text="페이스북으로 로그인하기"
+            android:paddingEnd="27dp"
+            android:gravity="center"
+            android:layout_marginTop="12dp"
+            app:layout_constraintTop_toBottomOf="@id/login_btn_kakao"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintWidth_default="percent"
+            android:background="@drawable/btn_login"
+            app:layout_constraintWidth_percent="0.631"
+            android:layout_height="48dp"/>
+
+    <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/login_btn_naver"
+            android:layout_width="0dp"
+            android:textSize="14sp"
+            android:textColor="#ffffff"
+            android:stateListAnimator="@null"
+            android:drawableStart="@drawable/ic_naver"
+            android:paddingStart="27dp"
+            android:text="네이버로 로그인하기"
+            android:paddingEnd="27dp"
+            android:gravity="center"
+            android:layout_marginTop="12dp"
+            app:layout_constraintTop_toBottomOf="@id/login_btn_facebook"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintWidth_default="percent"
+            android:background="@drawable/btn_login"
+            app:layout_constraintWidth_percent="0.631"
+            android:layout_height="48dp"/>
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
