+++ app/src/main/res/layout/activity_main.xml
@@ -3,7 +3,7 @@
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        tools:context=".MainActivity"
+        tools:context=".ui.MainActivity"
         android:id="@+id/main_drawer"
         android:layout_width="match_parent"
         tools:openDrawer="start"
@@ -97,7 +97,7 @@
                 tools:layout="@layout/main_bottom"
                 app:behavior_skipCollapsed="true"
                 app:layout_behavior="@string/bottom_sheet_behavior"
-                android:name="cooking.depromeet.github.com.bangguseokcookking_android.MainBottomFragment"/>
+                android:name="cooking.depromeet.github.com.bangguseokcookking_android.ui.MainBottomFragment"/>
 
 
     </androidx.coordinatorlayout.widget.CoordinatorLayout>
